% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_BART.R
\name{BART_site}
\alias{BART_site}
\title{Fit a BART Model with Study Site-Specific Random Intercepts}
\usage{
BART_site(
  df_train,
  Y_scaled,
  num_tree = 200,
  sim_num = 5000,
  df_test,
  num_site,
  factor_site
)
}
\arguments{
\item{df_train}{A data matrix containing the training predictors.}

\item{Y_scaled}{A numeric vector containing the scaled outcome values.}

\item{num_tree}{An integer specifying the number of trees in the BART ensemble.}

\item{sim_num}{An integer specifying the number of MCMC iterations.}

\item{df_test}{A data matrix containing the test predictors.}

\item{num_site}{An integer specifying the number of study sites.}

\item{factor_site}{A vector indicating the study site for each observation
in the training data.}
}
\value{
A list containing predicted values for the training and test data
and variable importance measures.
}
\description{
Fits a Bayesian Additive Regression Trees (BART) model with random
intercepts for study sites. This function is based on the SoftBart R
package and the BART framework described by Linero and Yang (2018).
}
