% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phewas_fun.R
\name{phewas_fun}
\alias{phewas_fun}
\title{Implement PheWAS}
\usage{
phewas_fun(
  target_phecode,
  covariates,
  genotypes,
  phecode_input,
  min_cases = 20
)
}
\arguments{
\item{target_phecode}{A data matrix containing subject IDs and a binary indicator for the target phecode.}

\item{covariates}{A matrix of covariates to adjust for in the PheWAS model.}

\item{genotypes}{A vector of PRS.}

\item{phecode_input}{The target phecode to be analyzed.}

\item{min_cases}{Minimum number of cases required to use logistic regression (\code{glm}).
If the number of cases is less than or equal to \code{min_cases}, Firth logistic regression (\code{logistf}) is used.}
}
\value{
A list containing the coefficient, p-val, SE, num. of cases, and associated phecode.
}
\description{
Performs a phenome-wide association study (PheWAS) using phecodes generated by the PheWAS package.
}
