% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_funcs_BART.R
\name{quantile_normalize}
\alias{quantile_normalize}
\title{Preprocess a Data Frame for BART}
\usage{
quantile_normalize(X, test_X)
}
\arguments{
\item{X}{A data frame.}

\item{test_X}{test data matrix.}
}
\description{
Apply quantile_normalize to the combined data matrix
}
