% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_funcs_BART.R
\name{quantile_normalize_bart}
\alias{quantile_normalize_bart}
\title{Quantile Normalization}
\usage{
quantile_normalize_bart(X)
}
\arguments{
\item{X}{A data matrix.}
}
\value{
A normalized data matrix.
}
\description{
Applies rank-based normalization to each column.
}
