% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_funcs_BART.R
\name{trank}
\alias{trank}
\title{Transform a Variable Using Rank-Based Normalization}
\usage{
trank(x)
}
\arguments{
\item{x}{A numeric vector to be transformed.}
}
\value{
A numeric vector with the transformed values.
}
\description{
Transforms a numeric vector using its unique ranks and scales the resulting values.
}
